TASK: Decouple PublishButton from AbstractButton as they diverged and are separate components
The `PublishButton` does not get a .5 opacity when disabled

... i almost created a `isVisuallyEnabled` flag to avoid the opacity.
