+++ packages/neos-ui/src/Containers/PrimaryToolbar/PublishDropDown/Components/AbstractButton/style.module.css
@@ -1,0 +1,18 @@
+.btn {
+    display: inline-block;
+    padding: 0 var(--spacing-Full);
+    height: var(--spacing-GoldenUnit);
+    line-height: var(--spacing-GoldenUnit);
+    border: 0;
+    text-align: center;
+    vertical-align: top;
+    cursor: pointer;
+}
+.btn--disabled {
+    opacity: .5;
+}
+.btn--notAllowed {
+    &[disabled] {
+        cursor: not-allowed;
+    }
+}
