+++ packages/neos-ui/src/Containers/PrimaryToolbar/PublishDropDown/index.js
@@ -10,6 +10,9 @@ import {translate} from '@neos-project/neos-ui-i18n';
 import {actions, selectors} from '@neos-project/neos-ui-redux-store';
 import {PublishingMode, PublishingScope} from '@neos-project/neos-ui-redux-store/src/CR/Publishing';
 import {neos} from '@neos-project/neos-ui-decorators';
+import {AbstractButton} from './Components/AbstractButton';
+import {PublishButton} from './Components/PublishButton';
+import style from './style.module.css';
 
 const {
     publishableNodesSelector,
@@ -18,8 +21,6 @@ const {
     personalWorkspaceNameSelector
 } = selectors.CR.Workspaces;
 
-import AbstractButton from './AbstractButton/index';
-import style from './style.module.css';
 
 @connect(state => ({
     isSaving: state?.ui?.remote?.isSaving,
@@ -90,7 +91,7 @@ export default class PublishDropDown extends PureComponent {
         const publishableNodesCount = publishableNodes ? publishableNodes.length : 0;
         return (
             <div id="neos-PublishDropDown" className={style.wrapper}>
-                <AbstractButton
+                <PublishButton
                     id="neos-PublishDropDown-Publish"
                     isEnabled={!isWorkspaceReadOnly && (canPublishLocally)}
                     isHighlighted={canPublishLocally || isSaving || isPublishing}
@@ -98,7 +99,7 @@ export default class PublishDropDown extends PureComponent {
                     >
                     {mainButton} {isWorkspaceReadOnly ? (<Icon icon="lock"/>) : ''}
                     {publishableNodesInDocumentCount > 0 && <Badge className={style.badge} label={String(publishableNodesInDocumentCount)}/>}
-                </AbstractButton>
+                </PublishButton>
 
                 <DropDown className={style.dropDown}>
                     {isPublishing || isSaving ? (
@@ -121,7 +122,6 @@ export default class PublishDropDown extends PureComponent {
                             <AbstractButton
                                 id="neos-PublishDropDown-PublishAll"
                                 isEnabled={!isWorkspaceReadOnly && canPublishGlobally}
-                                isHighlighted={false}
                                 onClick={this.handlePublishAllClick}
                                 >
                                 <div className={style.dropDown__iconWrapper}>
@@ -135,7 +135,6 @@ export default class PublishDropDown extends PureComponent {
                             <AbstractButton
                                 id="neos-PublishDropDown-Discard"
                                 isEnabled={canPublishLocally}
-                                isHighlighted={false}
                                 label="Discard"
                                 icon="ban"
                                 onClick={this.handleDiscardClick}
@@ -151,7 +150,6 @@ export default class PublishDropDown extends PureComponent {
                             <AbstractButton
                                 id="neos-PublishDropDown-DiscardAll"
                                 isEnabled={canPublishGlobally}
-                                isHighlighted={false}
                                 onClick={this.handleDiscardAllClick}
                                 >
                                 <div className={style.dropDown__iconWrapper}>
